Create the move in command and the move in action (#421)
* SQL migrations to create the mapping tables and populate them

* Add maps and java services to handle them

* Current work

* Command progress

* Rough first draft of the move in method. As is - we won't need a move in command

* I needed to add the entitlementId to be transfered to the move in. Also moved the file from core to web since that is the only place its used (so far)

* Rough draft of the complete command implementation

* Populated entitlementId, updated unit test

* Commit latest changes

* Refactoring, unit tests

* Removed unnecessary code from unit test

* Add unit tests for image, spec mappings as well as insert action

* Add the move out command

* Add migration for new action types

* Updated Guice injector mock in unit test

* Add unit tests for MoveIn endpoint, and a few bits of cleanup

* Simplify the move in command request

* Format the panotpa details a little better

* Fix a couple of errors in the SQL queries.
;

* Update web/src/main/java/com/godaddy/vps4/web/vm/PlatformMigrationResource.java


* Update web/src/main/java/com/godaddy/vps4/web/vm/PlatformMigrationResource.java


* Review comments

* Fist time getting move in to work. Now to clean up

* Change to just use the hfsImageName and specName in the moveOutInfo

* Update core/src/main/java/com/godaddy/vps4/vm/jdbc/JdbcVirtualMachineService.java


* A couple PR comment updates, and set the IP address valid until on move out.

* Fix unit tests

---------

+++ core/src/main/java/com/godaddy/vps4/move/VmMoveImageMapService.java
@@ -4,5 +4,5 @@ import com.godaddy.vps4.vm.ServerType;
 
 public interface VmMoveImageMapService {
 
-    VmMoveImageMap getVmMoveImageMap(int originalImageId, ServerType.Platform toPlatform);
+    VmMoveImageMap getVmMoveImageMap(long originalImageId, ServerType.Platform toPlatform);
 }
